Revert "move every matrix scan return to bool"
This reverts commit a291a9460b7c8e2f166bd36e403aa031584cb153.

+++ keyboards/bpiphany/pegasushoof/2015/matrix.c
@@ -92,7 +92,7 @@ uint8_t matrix_scan(void)
   debounce(matrix_debouncing, matrix, matrix_rows(), changed);
   matrix_scan_quantum();
 
-  return changed;
+  return (uint8_t)changed;
 }
 
 inline
