+++ keyboards/handwired/symmetric70_proto/matrix_fast/matrix.c
@@ -230,5 +230,5 @@ uint8_t matrix_scan(void) {
     MATRIX_DEBUG_SCAN_START();
     matrix_scan_quantum();
     MATRIX_DEBUG_SCAN_END();
-    return changed;
+    return (uint8_t)changed;
 }
