+++ keyboards/moon/matrix.c
@@ -205,5 +205,5 @@ uint8_t matrix_scan(void) {
 
   matrix_scan_quantum();
 
-  return changed;
+  return (uint8_t)changed;
 }
