+++ keyboards/ymdk/sp64/matrix.c
@@ -135,7 +135,7 @@ uint8_t matrix_scan(void)
       if (matrix_is_on(r, c)) xprintf("r:%d c:%d \n", r, c);
 #endif
 
-  return changed;
+  return (uint8_t)changed;
 }
 
 inline
